Add documentation, move libs into folders
