+++ src/game/map/map-generation/TerrainVer/LICENSE.md
@@ -1,0 +1,31 @@
+
+This project uses open-source libraries with the following licences:
+- https://github.com/josephg/noisejs, ISC_license
+- http://lodev.org/cgtutor/floodfill.html#Scanline_Floodfill_Algorithm_With_Stack, BSD-2-Clause licence
+- https://github.com/phoboslab/WebGLImageFilter, MIT licence
+- Google Blob Emojis, Apache License Version 2.0
+- https://github.com/phoboslab/js-hqx, LGPLv2.1 license, https://tldrlegal.com/license/gnu-lesser-general-public-license-v2.1-(lgpl-2.1)
+
+(All of these licences allow for commercial use)
+
+TerrainVer itself is licenced under the MIT License:
+
+Copyright (c) 2017 Julian Go, juliango202.com
+
+Permission is hereby granted, free of charge, to any person obtaining a copy
+of this software and associated documentation files (the "Software"), to deal
+in the Software without restriction, including without limitation the rights
+to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+copies of the Software, and to permit persons to whom the Software is
+furnished to do so, subject to the following conditions:
+
+The above copyright notice and this permission notice shall be included in all
+copies or substantial portions of the Software.
+
+THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+SOFTWARE.
