+++ src/game/map/map-generation/hxGeomAlgo/LICENSE.md
@@ -1,0 +1,64 @@
+## Credits
+
+**hxGeomAlgo** is based on the work of many developers and it wouldn't exist if it weren't for them. See the [CREDITS](CREDITS.md) file for a full list. 
+ 
+## License
+
+**hxGeomAlgo** is developed by Giuseppe Di Mauro (azrafe7) and all code is released under the MIT license unless specified otherwise.
+
+> The MIT License (MIT)
+> 
+> Copyright (c) 2013-2016 Giuseppe Di Mauro (azrafe7)
+> 
+> Permission is hereby granted, free of charge, to any person obtaining a copy
+> of this software and associated documentation files (the "Software"), to deal
+> in the Software without restriction, including without limitation the rights
+> to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+> copies of the Software, and to permit persons to whom the Software is
+> furnished to do so, subject to the following conditions:
+> 
+> The above copyright notice and this permission notice shall be included in all
+> copies or substantial portions of the Software.
+> 
+> THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+> IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+> FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+> AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+> LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+> OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+> SOFTWARE.
+
+The `Tess2` part of the library is released under the SGI Free Sofware License B v2.0:
+
+> SGI FREE SOFTWARE LICENSE B (Version 2.0, Sept. 18, 2008) 
+> 
+> Copyright (C) [dates of first publication] Silicon Graphics, Inc.
+> All Rights Reserved.
+> 
+> Permission is hereby granted, free of charge, to any person obtaining a copy
+> of this software and associated documentation files (the "Software"), to deal
+> in the Software without restriction, including without limitation the rights
+> to use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies
+> of the Software, and to permit persons to whom the Software is furnished to do so,
+> subject to the following conditions:
+> 
+> The above copyright notice including the dates of first publication and either this
+> permission notice or a reference to http://oss.sgi.com/projects/FreeB/ shall be
+> included in all copies or substantial portions of the Software. 
+> 
+> THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED,
+> INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A
+> PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL SILICON GRAPHICS, INC.
+> BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT,
+> TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE
+> OR OTHER DEALINGS IN THE SOFTWARE.
+> 
+> Except as contained in this notice, the name of Silicon Graphics, Inc. shall not
+> be used in advertising or otherwise to promote the sale, use or other dealings in
+> this Software without prior written authorization from Silicon Graphics, Inc.
+> 
+> Copyright (c) 1994 Eric Veach
+> 
+> Copyright (c) 2013 Mikko Mononen
+> 
+> Copyright (c) 2014 Giuseppe Di Mauro (azrafe7)
